Project specific plugins now get reloaded on project reload
When a project is unloaded all project specific plugins are shut down and the plugin is removed from the plugin manager.  When a project with that same plugin is loaded Glue attempts to load that plugin again, however since the plugin still exists in the AppDomain the `HandleAddingPLuginInDirectory()` method doesn't even try to load it (as it thinks it's already loaded).  This causes the plugin to be ignored.

This commit fixes it by handling plugins even if they are loaded in the AppDomain already.

The longer fix is to move away from the old AppDomain assembly loading and instead moving to the .net core `AseemblyLoadContext`.  `AssemblyLoadContext` supports unloading, which would allow for unloading the plugin on project change which would remove the need for this.