Initial loading and unloading of plugins.
While this works, it cannot be merged in due to https://github.com/dotnet/runtime/issues/1388.  Specifically, the Tiled plugin uses the `XmlSerializer` type, but the `XmlSerializer` tries to create dynamic assemblies in the default `AssemblyLoadContext`, which is not collectible.  Therefore you have a collectible assembly creating and loading a non-collectible assembly, which is invalid and thus an exception occurs.

So far this does not look like it will be fixed until .net 6 at the earliest.