Only project specific plugins should be unloadable
Some plugins throw exceptions when run under an unloadable/collectible assembly load context due to missing functionality in the .net core standard framework (specifically `XmlSerializer` usage).  Right now the main offendor of this is the Tiled plugin.  plugin

Since this functionality won't be fixed until .net 6 at the earliest, in order to support unloadable plugins we have to make the concession that global plugins are never unloadable (which seems reasonable) and only project specific plugins are loaded in an unloadable context.  This means that some plugins (like Tiled) will never be able to run at the project level until the .net runtime issues are resolved or Tiled is updated to not call the XmlSerializer.