Temporary project plugin directories
Currently Glue runs project plugins in the project's own `Plugins\` directory.  This causes plugin assemblies to be locked, which prevents you from installing new plugins (usually via nuget upgrades or downgrades).

In order to fix this, every time Glue loads a project it will create a temporary project specific plugin directory and run the plugins from that temporary directory.  This allows plugins to be dynamically changed, and on a project reload the old plugins will be unloaded and the new plugins loaded.