Added TimeManager.CurrentSystemTime to clean things up.
Wizard no longer sets the collision to be from Map if there is no Map created through the wizard.
Added Gum option to RemoveOrphanCustomCode.
Adding an entity list to the GameScreen through the quick actions no longer selects that list. Keeps the flow going.
Added wizard option to modify the collision type added to the player.