Instantiate PositionedObjectLists in containing element constructors
Run through WriteCodeForNamedObjectInitialize() for all named objects in
an element (entity or scene) for the constructor, excluding collision
relationships. Also update WriteCodeForNamedObjectInitialize() to check
if it is being called for emitting code into a constructor or not, since
we want to only emit instantiation code for contained objects in the
Initialize() method being generated, and not in constructors.