Cancel any tasks waiting on delay seconds when we are clearing tasks.
When moving from one screen to the other, time manager will clear all saved TCS objects.  However, the tasks that `DelaySeconds()` callers are waiting for still exist and will stay in the task scheduler forever.

This now sets the tasks to a cancelled state, and tells the synchronization context to ignore any `TaskCancellationException` that get raised by closing tasks.