Fixes #1412
Previously, the task manager would use a `foreach` to loop over the task queue via `GetConsumingEnumerable()`.  However, since this is a blocking collection, this locks the thread up.  This became an issue because we have functionality to explicitely make sure all async continuations are executed on this thread, but since the thread was blocked the task scheduler could not marshall the continuations onto it.

This commit fixes it by using `TryTake()`, and if no items exist we use `Task.Delay` to free the thread up to handle any continuations.