Remove PluginTabPage
PluginTab(viewmodel) had a handful of properties and logic that were just pass-throughs to the PluginTabPage control. These things can(should) just live on the viewmodel itself. This control was a blocker for me on my current task of theming/styling the app, as it's context menu was regenerated in the code-behind quite often and it's children being generated in the code behind didn't allow for a lot of wiggle room using a xaml template for the tab items.